+++ application/models/Tracker/Sites/WhiteCloudPavillion.php
@@ -1,5 +1,5 @@
 <?php declare(strict_types=1); defined('BASEPATH') OR exit('No direct script access allowed');
 
-class WhiteCloudPavillion extends Base_FoolSlide_Site_Model {
-	public $baseURL = 'https://archangelscans.com/free';
+class WhiteCloudPavillion extends Base_myMangaReaderCMS_Site_Model {
+	public $baseURL = 'https://whitecloudpavilion.com/manga/free';
 }