+++ public/userscripts/manga-tracker.meta.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         Manga Tracker
 // @namespace    https://github.com/DakuTree/userscripts
-// @version      1.8.11
+// @version      1.8.12
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // ==/UserScript==